ADB MONITOR 修正缓存命中率由于数值溢出问题

+++ src/adbmgrd/manager/adbmgr_init.sql
@@ -512,8 +512,8 @@ CREATE OR REPLACE FUNCTION pg_catalog.monitor_databasetps_func(in text, in times
 		RETURNS TABLE
 	(
 		recordtime timestamptz(0),
-		tps int,
-		qps int
+		tps bigint,
+		qps bigint
 	)
 	AS
 		$$
@@ -537,8 +537,8 @@ CREATE OR REPLACE FUNCTION pg_catalog.monitor_databasetps_func_by_time_period(db
 		RETURNS TABLE
 	(
 		recordtime timestamptz(0),
-		tps int,
-		qps int
+		tps bigint,
+		qps bigint
 	)
 	AS $$
 	SELECT monitor_databasetps_time::timestamptz(0) AS recordtime,
@@ -590,14 +590,14 @@ CREATE OR REPLACE FUNCTION pg_catalog.monitor_databasesummary_func(in name)
 		autovacuum		bool,
 		heaphitrate		numeric(18,2),
 		commitrate		numeric(18,2),
-		dbage			int,
-		connectnum		int,
-		standbydelay	int,
-		locksnum		int,
-		longtransnum	int,
-		idletransnum	int,
-		preparenum		int,
-		unusedindexnum	int
+		dbage			bigint,
+		connectnum		bigint,
+		standbydelay	bigint,
+		locksnum		bigint,
+		longtransnum	bigint,
+		idletransnum	bigint,
+		preparenum		bigint,
+		unusedindexnum	bigint
 	)
 	AS
 		$$
@@ -1014,7 +1014,6 @@ returns int
 as
 $$
     update pg_catalog.monitor_user set username=$2, usertel=$3, useremail=$4, usertitle=$5, usercompany=$6, userdesc=$7 where oid=$1 returning 0;
-
 $$
 LANGUAGE SQL
 VOLATILE
